+++ app/(navigatable)/(about)/about/components/AboutPolyverse.tsx
@@ -1,0 +1,10 @@
+const AboutPolyverse = () => (
+    <section>
+        <h2>About Polyverse</h2>
+        <p>Polyverse develops leading-edge AI-driven software development modernization tools and cybersecurity technology.</p>
+        <p>Sara, the AI Architect, is our newest product that uses AI to modernize software development and architecture. Sara is a powerful AI tool that helps developers and architects build better software faster.</p>
+        <p>Our Polymorphing and Boost Cybersecurity technologies are used by government and security-conscious organizations to mitigate against zero-day memory exploits. It is also embedded into devices, hardware and security solutions to provide the ultimate protection against hackers. Polyverse&apos;s Polymorphing technology is used by the U.S. Navy and a number of government agencies.</p>
+    </section>
+);
+
+export default AboutPolyverse;