+++ app/(navigatable)/(about)/about/components/DataRetention.tsx
@@ -1,0 +1,8 @@
+const DataRetention = () => (
+    <section>
+        <h2>Data Retention Policy</h2>
+        <p>All user data is not shared outside of the user&apos;s or organization&apos;s account. Any data tied to the user&apos;s project is contained within that project and can be deleted by deleting the project. The user&apos;s data is not used for training of any kind for other users outside of the user&apos;s organization.</p>
+    </section>
+);
+
+export default DataRetention;