+++ app/(navigatable)/(about)/about/components/Disclaimer.tsx
@@ -1,0 +1,10 @@
+import { preReleaseServiceDisclaimer } from "lib/productDescriptions";
+
+const Disclaimer = () => (
+    <section>
+      <h2>Disclaimer</h2>
+      <p>{preReleaseServiceDisclaimer}</p>
+    </section>
+  );
+  
+  export default Disclaimer;