+++ app/(navigatable)/(about)/about/components/FAQ.tsx
@@ -1,0 +1,18 @@
+const FAQ = () => (
+    <section>
+      <h2>Frequently Asked Questions</h2>
+      <dl>
+        <dt>What is Sara?</dt>
+        <dd>An AI Architectural Assistant.</dd>
+        
+        <dt>What is Boost?</dt>
+        <dd>AI Code Analysis services.</dd>
+        
+        <dt>Who is Polyverse?</dt>
+        <dd>A software engineering company with a long history of building developer and security products sold and licensed to government agencies and commercial companies.</dd>
+      </dl>
+    </section>
+  );
+  
+  export default FAQ;
+