+++ app/(navigatable)/(about)/about/page.tsx
@@ -1,0 +1,35 @@
+import { saraProductDescription } from 'lib/productDescriptions';
+import Layout from './layout';
+import Disclaimer from './components/Disclaimer';
+import ServicesUsed from './components/ServicesUsed';
+import DataRetention from './components/DataRetention';
+import FAQ from './components/FAQ';
+import AboutPolyverse from './components/AboutPolyverse';
+import Link from 'next/link';
+
+const AboutPage = () => (
+  <Layout>
+    <div className="flex flex-col items-center p-10 font-bold space-y-4">
+      <p className="text-2xl font-bold">About Sara AI by Polyverse</p>
+      <div className="w-1/2 border-t-2 border-blue-600 my-2"></div>
+    </div>
+    <p>{saraProductDescription}</p>
+    <p>The following information covers important information about Sara&apos;s technology, service availability, data retention policies, and common questions.</p>
+    <Disclaimer />
+    <ServicesUsed />
+    <DataRetention />
+    <FAQ />
+    <AboutPolyverse />
+    <div
+        id="footer"
+      >
+        <div className="text-center pt-24">
+          <Link href="https://www.polyverse.com">
+            <p>© 2024, Polyverse. All rights reserved.</p>
+          </Link>
+        </div>
+      </div>
+  </Layout>
+);
+
+export default AboutPage;