+++ app/(navigatable)/(orgs)/page.tsx
@@ -5,6 +5,7 @@ import { useRouter } from 'next/navigation'
 import { useAppContext } from 'lib/hooks/app-context'
 
 import OrgDashboard from './orgs/org-dashboard'
+import SaraLoading from 'components/sara-loading'
 
 const IndexPage = () => {
   const router = useRouter()
@@ -48,14 +49,10 @@ const IndexPage = () => {
     <div className="flex-1 p-10 text-2xl font-bold">
       <OrgDashboard orgs={orgs} />
       {isLoadingOrgs && (
-        <div className="flex justify-center items-center">
-          Sara is refreshing your account...
-        </div>
+        <SaraLoading message="Sara is refreshing your account..." />
       )}
       {isRedirectingToProjects && (
-        <div className="flex justify-center items-center pt-20">
-          Sara is refreshing your projects...
-        </div>
+        <SaraLoading message="Sara is refreshing your projects..." />
       )}
     </div>
   )